+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changelog
 ### 1.3.0 (16-11-2020)
 #### Improvements
-* Added configuration options for `--squash` and `--layers`. Can be configured in the `<build>` section of an image.
+* Added configuration options for `--squash`, `--squash-all` and `--layers`. Can be configured in the `<build>` section of an image.
 
 ### 1.2.2 (28-10-2020)
 #### Bugs